fix: reunited the beforesuite and aftersuite function.
